Support multiple Packages per Version
This implementation is not complete. The API isn't yet updated to support the new changes. Amazon S3 support also isn't updated. Finally, the migration added in this file needs to be modified to not just delete Version#file, but to also migrate existing data in that field to Package#file.